pyproject.toml
[tool.poetry]
name = "service-cert"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
uvloop = "^0.21.0"
alembic = "^1.13.3"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
email-validator = "^2.2.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = {version = "^2.0.36", extras = ["asyncio"]}
fastapi-jsonrpc = "^3.2.1"
uvicorn = "^0.33.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.2"
pre-commit = "^4.0.1"
polyfactory = "^2.18.1"
httpx = "^0.28.1"
dirty-equals = "^0.8.0"

[tool.pytest.ini_options]
testpaths = ["tests/"]
addopts = [
  "-p no:cacheprovider",
  "-p no:warnings",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"


[tool.ruff]
line-length = 100


[tool.ruff.lint]
ignore = [
  "F401",
  "F403",
]

[tool.ruff.format]
quote-style = "single"
